refactor(guardrails): own the rule descriptor and hash, drop responsecache import
GuardrailRuleDescriptor and ComputeGuardrailsHash were defined in
responsecache although they are guardrail-domain concepts, forcing the
guardrails package to depend on the cache layer. Move them into guardrails
(descriptor renamed to RuleDescriptor to avoid package stutter); the cache
keeps consuming the hash as an opaque context value, which was already the
only way it received it. No behavior change; hash tests move with the code.
