+++ internal/workflows/compiler_test.go
@@ -6,7 +6,6 @@ import (
 
 	"gomodel/internal/core"
 	"gomodel/internal/guardrails"
-	"gomodel/internal/responsecache"
 )
 
 func TestCompilerCompile_Guardrails(t *testing.T) {
@@ -15,7 +14,7 @@ func TestCompilerCompile_Guardrails(t *testing.T) {
 	if err != nil {
 		t.Fatalf("NewSystemPromptGuardrail() error = %v", err)
 	}
-	if err := registry.Register(rule, responsecache.GuardrailRuleDescriptor{
+	if err := registry.Register(rule, guardrails.RuleDescriptor{
 		Type:    "system_prompt",
 		Mode:    string(guardrails.SystemPromptInject),
 		Content: "be precise",
@@ -162,7 +161,7 @@ func TestCompilerCompile_WrapsBuildPipelineErrorsAsGatewayErrors(t *testing.T) {
 	if err != nil {
 		t.Fatalf("NewSystemPromptGuardrail() error = %v", err)
 	}
-	if err := registry.Register(rule, responsecache.GuardrailRuleDescriptor{
+	if err := registry.Register(rule, guardrails.RuleDescriptor{
 		Name:    "present",
 		Type:    "system_prompt",
 		Mode:    string(guardrails.SystemPromptInject),
